Refactor code to fix cmake compile issues
