+++ LTexture.h
@@ -1,6 +1,7 @@
 #pragma once
-#include "libs.h"
-#include <SDL_ttf.h>
+#include <iostream>
+#include <SDL2/SDL_ttf.h>
+#include <SDL2/SDL_image.h>
 /**/
 class LTexture
 {
