+++ Stroka.h
